+++ src/tespy/connections/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8
+
+from .bus import Bus  # noqa: F401
+from .connection import Connection  # noqa: F401
+from .ref import Ref  # noqa: F401
