+++ src/tespy/networks/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8
 from .network_reader import load_network  # noqa: F401
-from .networks import Network  # noqa: F401
+from .network import Network  # noqa: F401
